+++ src/superface-test.ts
@@ -6,8 +6,6 @@ import {
   Result,
   SuperfaceClient,
 } from '@superfaceai/one-sdk';
-import { Primitive } from '@superfaceai/one-sdk/dist/internal/interpreter/variables';
-import { createHash } from 'crypto';
 import createDebug from 'debug';
 import {
   activate as activateNock,
@@ -21,6 +19,7 @@ import {
 import { join as joinPath } from 'path';
 
 import {
+  InputVariables,
   ProcessingFunction,
   RecordingDefinitions,
   RecordingProcessOptions,
@@ -39,6 +38,7 @@ import {
 } from './common/format';
 import { exists, readFileQuiet } from './common/io';
 import { writeRecordings } from './common/output-stream';
+import { IGenerator } from './generate-hash';
 import {
   NockConfig,
   SuperfaceTestConfigPayload,
@@ -50,6 +50,7 @@ import {
   assertsDefinitionsAreNotStrings,
   assertsPreparedConfig,
   checkSensitiveInformation,
+  getGenerator,
   getProfileId,
   getSuperJson,
   isProfileProviderLocal,
@@ -58,6 +59,8 @@ import {
 } from './superface-test.utils';
 
 const debug = createDebug('superface:testing');
+const debugSetup = createDebug('superface:testing:setup');
+const debugHashing = createDebug('superface:testing:hash');
 
 export class SuperfaceTest {
   private sfConfig: SuperfaceTestConfigPayload;
@@ -65,10 +68,12 @@ export class SuperfaceTest {
   private nockConfig?: NockConfig;
   private fixturesPath?: string;
   private recordingPath?: string;
+  private generator: IGenerator;
 
   constructor(sfConfig?: SuperfaceTestConfigPayload, nockConfig?: NockConfig) {
     this.sfConfig = sfConfig ?? {};
     this.nockConfig = nockConfig;
+    this.generator = getGenerator(sfConfig?.testInstance);
 
     this.setupFixturesPath();
   }
@@ -93,15 +98,15 @@ export class SuperfaceTest {
         this.sfConfig.provider.configuration
       );
 
-    const hash = createHash('md5')
-      .update(JSON.stringify(testCase.input))
-      .digest('hex');
+    // Create a hash for access to recording files
+    const { input, testName } = testCase;
+    const hash = this.generator.hash({ input, testName });
 
-    debug('Created hash based on input value:', hash);
+    debugHashing('Created hash:', hash);
 
     this.setupRecordingPath(getFixtureName(this.sfConfig), hash);
 
-    // parse env variable and check if test should be recorded
+    // Parse env variable and check if test should be recorded
     const record = matchWildCard(this.sfConfig, process.env.SUPERFACE_LIVE_API);
     const processRecordings = options?.processRecordings ?? true;
     const inputVariables = searchValues(testCase.input, options?.hideInput);
@@ -115,7 +120,8 @@ export class SuperfaceTest {
 
     let result: Result<unknown, PerformError>;
     try {
-      result = await this.sfConfig.useCase.perform(testCase.input, {
+      // Run perform method on specified configuration
+      result = await this.sfConfig.useCase.perform(input, {
         provider: this.sfConfig.provider,
       });
 
@@ -161,7 +167,7 @@ export class SuperfaceTest {
   private async startRecording(
     record: boolean,
     processRecordings: boolean,
-    inputVariables?: Record<string, Primitive>,
+    inputVariables?: InputVariables,
     beforeRecordingLoad?: ProcessingFunction
   ): Promise<void> {
     if (!this.recordingPath) {
@@ -247,7 +253,7 @@ export class SuperfaceTest {
   private async endRecording(
     record: boolean,
     processRecordings: boolean,
-    inputVariables?: Record<string, Primitive>,
+    inputVariables?: InputVariables,
     beforeRecordingSave?: ProcessingFunction
   ): Promise<void> {
     if (!this.recordingPath) {
@@ -336,7 +342,7 @@ export class SuperfaceTest {
       this.fixturesPath = path ?? joinPath(process.cwd(), 'nock');
     }
 
-    debug('Prepare path to recording fixtures:', this.fixturesPath);
+    debugSetup('Prepare path to recording fixtures:', this.fixturesPath);
   }
 
   /**
@@ -353,7 +359,7 @@ export class SuperfaceTest {
       `${this.nockConfig?.fixture ?? 'recording'}-${inputHash}.json`
     );
 
-    debug('Prepare path to recording:', this.recordingPath);
+    debugSetup('Prepare path to recording:', this.recordingPath);
   }
 
   /**
@@ -376,7 +382,7 @@ export class SuperfaceTest {
       this.sfConfig.useCase = payload.useCase;
     }
 
-    debug('Superface configuration prepared:', this.sfConfig);
+    debugSetup('Superface configuration prepared:', this.sfConfig);
   }
 
   /**
@@ -387,7 +393,7 @@ export class SuperfaceTest {
     if (!this.sfConfig.client) {
       this.sfConfig.client = new SuperfaceClient();
 
-      debug('Superface client initialized:', this.sfConfig.client);
+      debugSetup('Superface client initialized:', this.sfConfig.client);
     }
 
     if (typeof this.sfConfig.profile === 'string') {
@@ -395,7 +401,7 @@ export class SuperfaceTest {
         this.sfConfig.profile
       );
 
-      debug('Superface Profile transformed:', this.sfConfig.profile);
+      debugSetup('Superface Profile transformed:', this.sfConfig.profile);
     }
 
     if (typeof this.sfConfig.provider === 'string') {
@@ -403,7 +409,7 @@ export class SuperfaceTest {
         this.sfConfig.provider
       );
 
-      debug('Superface Provider transformed:', this.sfConfig.provider);
+      debugSetup('Superface Provider transformed:', this.sfConfig.provider);
     }
 
     if (typeof this.sfConfig.useCase === 'string') {
@@ -415,7 +421,7 @@ export class SuperfaceTest {
         this.sfConfig.useCase
       );
 
-      debug('Superface UseCase transformed:', this.sfConfig.useCase);
+      debugSetup('Superface UseCase transformed:', this.sfConfig.useCase);
     }
   }